Remove Swagger v1
Removes all files to do with the v1 client. However, it does leave the
'version client', which doesn't appear to come from the v1 swagger, in
fact it's not clear where that's come from.

Adds running the tests to the circle config.

Removes the notion of 'callURL' from the provider chain as well, as we
have migrated over to the 'invoke' method, where the urls are provided
by the fn data.