Instance-Principal provider
OCI instances can take advantage of the instance principal authn
provider.

This provider will authenticate as the instance; that instance will
require the appropriate rights granting to it.

It defaults to picking a functions endpoint that is in the same
region as the instance.

(You'll need an explicit
    api-url: ""
 in the config to use this)

It defaults to using the compartment ID that the instance belongs to.