OCI shim: handle case where REGION is not present in OCI config
There was an issue with the new 'OCI shim' provider code, whereby, despite api-url being set in the Fn context, we still required REGION to be set in the OCI config. This is incorrect behaviour as we should require only api-url. However, OCI SDK client construction requires a non-blank region in the config provider, despite the fact that we override endpoint with api-url, so we have to work around that here.