Allow clients to specify content type
At the moment the signer always overrides the content type of the
request to be `application/json`. There's no need for this, and it
means that there's no way to override the header when necessary.

This change checks to see if the content type header is set, and if
the header is set then it's left as-is.